getting rid of useEffect
